Move object ref tests to separate file
